WIP Yet more moving things around; moved repository management from _data to commands and removed splitgraph.constants
