+++ benchmarking/commit_chain_test.py
@@ -3,8 +3,6 @@ os.environ['SG_CONFIG_FILE'] = 'test/resources/.sgconfig'
 
 from splitgraph.connection import get_connection, serialize_connection_string, make_conn
 from minio import Minio
-from splitgraph.config.repo_lookups import get_remote_connection_params
-
 
 from datetime import datetime
 from random import getrandbits, randrange
@@ -12,8 +10,8 @@ from random import getrandbits, randrange
 from psycopg2.extras import execute_batch
 
 from splitgraph.commands import *
-from splitgraph import to_repository, init, unmount
-from hooks.s3 import S3_HOST, S3_PORT, S3_ACCESS_KEY, S3_SECRET_KEY
+from splitgraph import to_repository, init, unmount, get_remote_connection_params
+from splitgraph.hooks.s3 import S3_HOST, S3_PORT, S3_ACCESS_KEY, S3_SECRET_KEY
 from splitgraph.commands.tagging import get_current_head
 
 MOUNTPOINT = to_repository("splitgraph_benchmark")
