+++ splitgraph/commandline.py
@@ -8,7 +8,6 @@ import click
 from psycopg2 import ProgrammingError
 
 import splitgraph as sg
-import splitgraph.commands.repository
 
 
 def _commit_connection(result):
@@ -26,7 +25,7 @@ def cli():
 
 
 @click.command(name='status')
-@click.argument('repository', required=False, type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', required=False, type=sg.to_repository)
 def status_c(repository):
     if repository is None:
         repositories = sg.get_current_repositories()
@@ -51,7 +50,7 @@ def status_c(repository):
 
 
 @click.command(name='log')
-@click.argument('repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', type=sg.to_repository)
 @click.option('-t', '--tree', is_flag=True)
 def log_c(repository, tree):
     if tree:
@@ -68,7 +67,7 @@ def log_c(repository, tree):
 @click.command(name='diff')
 @click.option('-v', '--verbose', default=False, is_flag=True)
 @click.option('-t', '--table-name')
-@click.argument('repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', type=sg.to_repository)
 @click.argument('tag_or_hash_1', required=False)
 @click.argument('tag_or_hash_2', required=False)
 def diff_c(verbose, table_name, repository, tag_or_hash_1, tag_or_hash_2):
@@ -154,13 +153,13 @@ def mount_c(schema, connection, handler, handler_options):
 
 
 @click.command(name='unmount')
-@click.argument('repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', type=sg.to_repository)
 def unmount_c(repository):
-    splitgraph.commands.repository.unmount(repository)
+    sg.unmount(repository)
 
 
 @click.command(name='checkout')
-@click.argument('repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', type=sg.to_repository)
 @click.argument('snapshot_or_tag')
 def checkout_c(repository, snapshot_or_tag):
     snapshot = sg.tag_or_hash_to_actual_hash(repository, snapshot_or_tag)
@@ -169,7 +168,7 @@ def checkout_c(repository, snapshot_or_tag):
 
 
 @click.command(name='commit')
-@click.argument('repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', type=sg.to_repository)
 @click.option('-h', '--commit-hash')
 @click.option('-s', '--include-snap', default=False, is_flag=True)
 @click.option('-m', '--message')
@@ -183,7 +182,7 @@ def commit_c(repository, commit_hash, include_snap, message):
 
 
 @click.command(name='show')
-@click.argument('repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', type=sg.to_repository)
 @click.argument('commit_tag_or_hash')
 @click.option('-v', '--verbose', default=False, is_flag=True)
 def show_c(repository, commit_tag_or_hash, verbose):
@@ -214,7 +213,7 @@ def show_c(repository, commit_tag_or_hash, verbose):
 @click.argument('splitfile', type=click.File('r'))
 @click.option('-a', '--splitfile-args', multiple=True, type=(str, str))
 @click.option('-o', '--output-repository', help='Repository to store the result in.',
-              type=splitgraph.commands.repository.to_repository)
+              type=sg.to_repository)
 def file_c(splitfile, splitfile_args, output_repository):
     splitfile_args = {k: v for k, v in splitfile_args}
     print("Executing Splitfile %s with arguments %r" % (splitfile.name, splitfile_args))
@@ -237,14 +236,14 @@ def sql_c(sql, show_all):
 
 
 @click.command(name='init')
-@click.argument('repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', type=sg.to_repository)
 def init_c(repository):
-    splitgraph.commands.repository.init(repository)
+    sg.init(repository)
     print("Initialized empty repository %s" % str(repository))
 
 
 @click.command(name='pull')
-@click.argument('repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', type=sg.to_repository)
 @click.argument('remote', default='origin')
 @click.option('-d', '--download-all', help='Download all objects immediately instead on checkout.')
 def pull_c(repository, remote, download_all):
@@ -252,8 +251,8 @@ def pull_c(repository, remote, download_all):
 
 
 @click.command(name='clone')
-@click.argument('remote_repository', type=splitgraph.commands.repository.to_repository)
-@click.argument('local_repository', required=False, type=splitgraph.commands.repository.to_repository)
+@click.argument('remote_repository', type=sg.to_repository)
+@click.argument('local_repository', required=False, type=sg.to_repository)
 @click.option('-d', '--download-all', help='Download all objects immediately instead on checkout.',
               default=False, is_flag=True)
 def clone_c(remote_repository, local_repository, download_all):
@@ -261,9 +260,9 @@ def clone_c(remote_repository, local_repository, download_all):
 
 
 @click.command(name='push')
-@click.argument('repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', type=sg.to_repository)
 @click.argument('remote', default='origin')  # origin (must be specified in the config or remembered by the repo)
-@click.argument('remote_repository', required=False, type=splitgraph.commands.repository.to_repository)
+@click.argument('remote_repository', required=False, type=sg.to_repository)
 @click.option('-h', '--upload-handler', help='Where to upload objects (FILE or DB for the remote itself)', default='DB')
 @click.option('-o', '--upload-handler-options', help="""For FILE, e.g. '{"path": /mnt/sgobjects}'""", default="{}")
 def push_c(repository, remote, remote_repository, upload_handler, upload_handler_options):
@@ -280,7 +279,7 @@ def push_c(repository, remote, remote_repository, upload_handler, upload_handler
 
 
 @click.command(name='tag')
-@click.argument('repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', type=sg.to_repository)
 @click.option('-i', '--image')
 @click.argument('tag', required=False)
 @click.option('-f', '--force', required=False, is_flag=True)
@@ -311,9 +310,9 @@ def tag_c(repository, image, tag, force):
 
 
 @click.command(name='import')
-@click.argument('repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', type=sg.to_repository)
 @click.argument('table_or_query')
-@click.argument('target_repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('target_repository', type=sg.to_repository)
 @click.argument('target_table', required=False)
 @click.argument('image', required=False)
 @click.option('-q', '--is-query', is_flag=True, default=False)
@@ -344,7 +343,7 @@ def cleanup_c():
 
 
 @click.command(name='provenance')
-@click.argument('repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', type=sg.to_repository)
 @click.argument('snapshot_or_tag')
 @click.option('-f', '--full', required=False, is_flag=True, help='Recreate the Splitfile used to create this image')
 @click.option('-e', '--error-on-end', required=False, default=True, is_flag=True,
@@ -363,10 +362,10 @@ def provenance_c(repository, snapshot_or_tag, full, error_on_end):
 
 
 @click.command(name='rerun')
-@click.argument('repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', type=sg.to_repository)
 @click.argument('snapshot_or_tag')
 @click.option('-u', '--update', is_flag=True, help='Rederive the image against the latest version of all dependencies.')
-@click.option('-i', '--repo-image', multiple=True, type=(splitgraph.commands.repository.to_repository, str))
+@click.option('-i', '--repo-image', multiple=True, type=(sg.to_repository, str))
 def rerun_c(repository, snapshot_or_tag, update, repo_image):
     snapshot = sg.tag_or_hash_to_actual_hash(repository, snapshot_or_tag)
 
@@ -385,7 +384,7 @@ def rerun_c(repository, snapshot_or_tag, update, repo_image):
 
 
 @click.command(name='publish')
-@click.argument('repository', type=splitgraph.commands.repository.to_repository)
+@click.argument('repository', type=sg.to_repository)
 @click.argument('tag')
 @click.option('-r', '--readme', type=click.File('r'))
 @click.option('--skip-provenance', is_flag=True, help='Don''t include provenance in the published information.')
