+++ splitgraph/commands/__init__.py
@@ -3,10 +3,11 @@ from splitgraph.commands.commit import commit
 from splitgraph.commands.diff import diff
 from splitgraph.commands.importing import import_tables
 from splitgraph.commands.info import get_image, get_canonical_image_id, get_parent_children, get_tables_at, get_table
-from splitgraph.commands.misc import get_log, cleanup_objects
+from splitgraph.commands.misc import get_log, cleanup_objects, init, unmount
 from splitgraph.commands.mounting import mount
 from splitgraph.commands.provenance import provenance, image_hash_to_splitfile
 from splitgraph.commands.publish import publish
 from splitgraph.commands.push_pull import push, pull, clone
-from splitgraph.commands.repository import Repository, to_repository, init, unmount
+from splitgraph.commands.repository import Repository, to_repository, repository_exists, \
+    get_remote_connection_params, get_remote_for, get_current_repositories
 from splitgraph.commands.tagging import get_current_head, get_all_hashes_tags, set_tag, tag_or_hash_to_actual_hash
