+++ splitgraph/commands/importing.py
@@ -2,13 +2,14 @@ from random import getrandbits
 
 from psycopg2.sql import Identifier, SQL
 
-from splitgraph import Repository, unmount
 from splitgraph._data.images import add_new_image
 from splitgraph._data.objects import register_table, register_object
 from splitgraph.commands._objects.utils import get_random_object_id
 from splitgraph.commands.checkout import materialize_table, checkout
 from splitgraph.commands.info import get_tables_at, get_table
+from splitgraph.commands.misc import unmount
 from splitgraph.commands.push_pull import clone
+from splitgraph.commands.repository import Repository
 from splitgraph.commands.tagging import get_current_head
 from splitgraph.config import SPLITGRAPH_META_SCHEMA
 from splitgraph.connection import get_connection
