+++ splitgraph/commands/misc.py
@@ -1,8 +1,10 @@
 from psycopg2.sql import SQL, Identifier
 
-from splitgraph._data.common import META_TABLES
+from splitgraph._data.common import META_TABLES, ensure_metadata_schema
 from splitgraph._data.objects import get_object_meta
+from splitgraph.commands._pg_audit import manage_audit, discard_pending_changes
 from splitgraph.commands.info import get_image, get_table
+from splitgraph.commands.repository import _register_repository, _unregister_repository
 from splitgraph.config import SPLITGRAPH_META_SCHEMA
 from splitgraph.connection import get_connection
 from splitgraph.pg_utils import pg_table_exists
@@ -99,3 +101,40 @@ def delete_objects(objects):
         with get_connection().cursor() as cur:
             cur.execute(SQL(";").join(SQL("DROP TABLE IF EXISTS {}.{}").format(Identifier(SPLITGRAPH_META_SCHEMA),
                                                                                Identifier(d)) for d in objects))
+
+
+@manage_audit
+def init(repository):
+    """
+    Initializes an empty repo with an initial commit (hash 0000...)
+
+    :param repository: Repository to create. Must not exist locally.
+    """
+    with get_connection().cursor() as cur:
+        cur.execute(SQL("CREATE SCHEMA {}").format(Identifier(repository.to_schema())))
+    image_hash = '0' * 64
+    _register_repository(repository, image_hash, tables=[], table_object_ids=[])
+
+
+def unmount(repository):
+    """
+    Discards all changes to a given repository and all of its history, deleting the physical Postgres schema.
+    Doesn't delete any cached physical objects.
+
+    :param repository: Repository to unmount.
+    :return:
+    """
+    # Make sure to discard changes to this repository if they exist, otherwise they might
+    # be applied/recorded if a new repository with the same name appears.
+    ensure_metadata_schema()
+    discard_pending_changes(repository.to_schema())
+    conn = get_connection()
+
+    with conn.cursor() as cur:
+        cur.execute(SQL("DROP SCHEMA IF EXISTS {} CASCADE").format(Identifier(repository.to_schema())))
+        # Drop server too if it exists (could have been a non-foreign repository)
+        cur.execute(SQL("DROP SERVER IF EXISTS {} CASCADE").format(Identifier(repository.to_schema() + '_server')))
+
+    # Currently we just discard all history info about the mounted schema
+    _unregister_repository(repository)
+    conn.commit()
