+++ test/splitgraph/splitfile/test_execution.py
@@ -2,12 +2,12 @@ import pytest
 
 from splitgraph import to_repository as R, unmount
 from splitgraph._data.images import get_all_images_parents
-from splitgraph._data.misc import get_current_repositories
 from splitgraph._data.objects import get_existing_objects, get_downloaded_objects, get_external_object_locations, \
     get_object_for_table
 from splitgraph.commands import checkout, commit, push, clone, get_log
 from splitgraph.commands.info import get_image, get_tables_at
 from splitgraph.commands.misc import cleanup_objects
+from splitgraph.commands.repository import get_current_repositories
 from splitgraph.commands.tagging import get_current_head
 from splitgraph.connection import override_driver_connection
 from splitgraph.exceptions import SplitGraphException
