+++ test/splitgraph/test_rls.py
@@ -3,16 +3,15 @@ import tempfile
 import pytest
 from psycopg2._psycopg import ProgrammingError
 
-from splitgraph import Repository
+from splitgraph import Repository, get_remote_connection_params
 from splitgraph._data.common import toggle_registry_rls
 from splitgraph._data.images import get_all_images_parents
-from splitgraph._data.misc import unregister_repository
 from splitgraph._data.objects import register_object_locations
 from splitgraph.commands import clone, push, commit, checkout
 from splitgraph.commands.info import get_table
 from splitgraph.commands.publish import publish
+from splitgraph.commands.repository import _unregister_repository
 from splitgraph.commands.tagging import get_tagged_id, set_tag
-from splitgraph.config.repo_lookups import get_remote_connection_params
 from splitgraph.connection import override_driver_connection, serialize_connection_string
 from splitgraph.registry_meta_handler import get_published_info, unpublish_repository
 from test.splitgraph.conftest import PG_MNT, add_multitag_dataset_to_remote_driver
@@ -64,7 +63,7 @@ def test_rls_push_own_delete_own(empty_pg_conn, unprivileged_conn_string, unpriv
 
     # Test we can delete our own repo once we've pushed it
     with override_driver_connection(unprivileged_remote_conn):
-        unregister_repository(target_repo, is_remote=True)
+        _unregister_repository(target_repo, is_remote=True)
     assert len(get_all_images_parents(target_repo)) == 0
 
 
@@ -85,7 +84,7 @@ def test_rls_delete_others(unprivileged_remote_conn):
     # RLS doesn't actually raise an error for this, since it just appends the policy qualifier to the query.
     # Hence in this case this simply does nothing (the rows in "test" namespace aren't available for deletion).
     with override_driver_connection(unprivileged_remote_conn):
-        unregister_repository(PG_MNT, is_remote=True)
+        _unregister_repository(PG_MNT, is_remote=True)
         # Check that the policy worked by verifying that the repository still exists on the remote.
         assert len(get_all_images_parents(PG_MNT)) > 0
